requirements.txt
Flask==2.2.2
gunicorn==20.1.0
pandas==1.5.0
Werkzeug==2.2.2
tabula-py==2.7.0
